requirements.txt
altgraph==0.17
auto-py-to-exe==2.9.0
bidict==0.21.2
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dataclasses==0.8
EasyProcess==0.3
Eel==0.12.4
entrypoint2==0.2.4
Flask==2.0.1
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
idna==2.10
importlib-metadata==4.5.0
importlib-resources==5.1.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
MouseInfo==0.1.3
mss==6.1.0
pefile==2021.5.24
Pillow==8.2.0
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyMsgBox==1.0.9
PyMuPDF==1.18.14
pyperclip==1.8.2
PyRect==0.1.4
pyscreenshot==3.0
PyScreeze==0.1.27
python-engineio==4.2.0
python-socketio==5.3.0
pythonnet==2.5.2
PyTweening==1.0.3
pywebview==3.4
pywin32-ctypes==0.2.0
requests==2.25.1
SQLAlchemy==1.4.17
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
whichcraft==0.6.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
